Add startsAtBottom prop to shared chart viewport
When this prop is true the viewport starts from the bottom, meaning
the user will need to scroll the view upwards in order to see more of
the graph.  It will also make sure the graph is anchored at the bottom
when the viewport height is larger than the height of the content.

This is useful for rendering a flame graph.